+++ Scraps.GUI/Services/Updater/Models/LatestRelease.cs
@@ -18,88 +18,158 @@
 
 namespace Scraps.GUI.Services.Updater.Models;
 
-/**
- * TODO: Fix name rule violation
- */
 public class LatestRelease
 {
-    public string url { get; set; }
-    public string assets_url { get; set; }
-    public string upload_url { get; set; }
-    public string html_url { get; set; }
-    public int id { get; set; }
-    public string node_id { get; set; }
-    public string tag_name { get; set; }
-    public string target_commitish { get; set; }
-    public string name { get; set; }
-    public bool draft { get; set; }
-    public Author author { get; set; }
-    public bool prerelease { get; set; }
-    public DateTime created_at { get; set; }
-    public DateTime published_at { get; set; }
-    public Asset[] assets { get; set; }
-    public string tarball_url { get; set; }
-    public string zipball_url { get; set; }
-    public string body { get; set; }
+    [JsonPropertyName("url")]
+    public string Url { get; set; }
+
+    [JsonPropertyName("assets_url")]
+    public string AssetsUrl { get; set; }
+
+    [JsonPropertyName("upload_url")]
+    public string UploadUrl { get; set; }
+
+    [JsonPropertyName("html_url")]
+    public string HtmlUrl { get; set; }
+
+    [JsonPropertyName("id")]
+    public int Id { get; set; }
+
+    [JsonPropertyName("node_id")]
+    public string NodeId { get; set; }
+
+    [JsonPropertyName("tag_name")]
+    public string TagName { get; set; }
+
+    [JsonPropertyName("target_commitish")]
+    public string TargetComitish { get; set; }
+
+    [JsonPropertyName("name")]
+    public string Name { get; set; }
+
+    [JsonPropertyName("draft")]
+    public bool Draft { get; set; }
+
+    [JsonPropertyName("author")]
+    public Author Author { get; set; }
+
+    [JsonPropertyName("prerelease")]
+    public bool Prerelease { get; set; }
+
+    [JsonPropertyName("created_at")]
+    public DateTime CreatedAt { get; set; }
+
+    [JsonPropertyName("published_at")]
+    public DateTime PublishedAt { get; set; }
+
+    [JsonPropertyName("assets")]
+    public Asset[] Assets { get; set; }
+
+    [JsonPropertyName("tarball_url")]
+    public string TarballUrl { get; set; }
+
+    [JsonPropertyName("zipball_url")]
+    public string ZipballUrl { get; set; }
+
+    [JsonPropertyName("body")]
+    public string Body { get; set; }
 }
 
 public class Author
 {
-    public string login { get; set; }
-    public int id { get; set; }
-    public string node_id { get; set; }
-    public string avatar_url { get; set; }
-    public string gravatar_id { get; set; }
-    public string url { get; set; }
-    public string html_url { get; set; }
-    public string followers_url { get; set; }
-    public string following_url { get; set; }
-    public string gists_url { get; set; }
-    public string starred_url { get; set; }
-    public string subscriptions_url { get; set; }
-    public string organizations_url { get; set; }
-    public string repos_url { get; set; }
-    public string events_url { get; set; }
-    public string received_events_url { get; set; }
-    public string type { get; set; }
-    public bool site_admin { get; set; }
+    [JsonPropertyName("login")]
+    public string Login { get; set; }
+
+    [JsonPropertyName("id")]
+    public int Id { get; set; }
+
+    [JsonPropertyName("node_id")]
+    public string NodeId { get; set; }
+
+    [JsonPropertyName("avatar_url")]
+    public string AvatarUrl { get; set; }
+
+    [JsonPropertyName("gravatar_id")]
+    public string GravatarId { get; set; }
+
+    [JsonPropertyName("url")]
+    public string Url { get; set; }
+
+    [JsonPropertyName("html_url")]
+    public string HtmlUrl { get; set; }
+
+    [JsonPropertyName("followers_url")]
+    public string FollowersUrl { get; set; }
+
+    [JsonPropertyName("following_url")]
+    public string FollowingUrl { get; set; }
+
+    [JsonPropertyName("gists_url")]
+    public string GistsUrl { get; set; }
+
+    [JsonPropertyName("starred_url")]
+    public string StarredUrl { get; set; }
+
+    [JsonPropertyName("subscriptions_url")]
+    public string SubscriptionsUrl { get; set; }
+
+    [JsonPropertyName("organizations_url")]
+    public string OrganizationsUrl { get; set; }
+
+    [JsonPropertyName("repos_url")]
+    public string ReposUrl { get; set; }
+
+    [JsonPropertyName("events_url")]
+    public string EventsUrl { get; set; }
+
+    [JsonPropertyName("received_events_url")]
+    public string ReceivedEventsUrl { get; set; }
+
+    [JsonPropertyName("type")]
+    public string Type { get; set; }
+
+    [JsonPropertyName("site_admin")]
+    public bool SiteAdmin { get; set; }
 }
 
 public class Asset
 {
-    public string url { get; set; }
-    public int id { get; set; }
-    public string node_id { get; set; }
-    public string name { get; set; }
-    public object label { get; set; }
-    public Uploader uploader { get; set; }
-    public string content_type { get; set; }
-    public string state { get; set; }
-    public int size { get; set; }
-    public int download_count { get; set; }
-    public DateTime created_at { get; set; }
-    public DateTime updated_at { get; set; }
-    public string browser_download_url { get; set; }
-}
+    [JsonPropertyName("url")]
+    public string Url { get; set; }
 
-public class Uploader
-{
-    public string login { get; set; }
-    public int id { get; set; }
-    public string node_id { get; set; }
-    public string avatar_url { get; set; }
-    public string gravatar_id { get; set; }
-    public string url { get; set; }
-    public string html_url { get; set; }
-    public string followers_url { get; set; }
-    public string following_url { get; set; }
-    public string gists_url { get; set; }
-    public string starred_url { get; set; }
-    public string subscriptions_url { get; set; }
-    public string organizations_url { get; set; }
-    public string repos_url { get; set; }
-    public string events_url { get; set; }
-    public string received_events_url { get; set; }
-    public string type { get; set; }
-    public bool site_admin { get; set; }
+    [JsonPropertyName("id")]
+    public int Id { get; set; }
+
+    [JsonPropertyName("node_id")]
+    public string NodeId { get; set; }
+
+    [JsonPropertyName("name")]
+    public string name { get; set; }
+
+    [JsonPropertyName("label")]
+    public object Label { get; set; }
+
+    [JsonPropertyName("uploader")]
+    public Author Uploader { get; set; }
+
+    [JsonPropertyName("content_type")]
+    public string ContentType { get; set; }
+
+    [JsonPropertyName("state")]
+    public string State { get; set; }
+
+    [JsonPropertyName("size")]
+    public int Size { get; set; }
+
+    [JsonPropertyName("download_count")]
+    public int DownloadCount { get; set; }
+
+    [JsonPropertyName("created_at")]
+    public DateTime CreatedAt { get; set; }
+
+    [JsonPropertyName("updated_at")]
+    public DateTime UpdatedAt { get; set; }
+
+    [JsonPropertyName("browser_download_url")]
+    public string BrowserDownloadUrl { get; set; }
 }