+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 4.10.0.0
+
+The .NET 7 Desktop Runtime is now required to run Scraps.
+
+- General code cleanup
+- Now uses AngleSharp for HTML parsing
+- Logged-in username is now logged on runner start
+- HTTP requests are logged to debug
+- Updated dependencies
+
 ## 4.9.1.0
 
 - Re-added cookie value length validation when saving settings